requirements.txt
flask>=2.3.0
openai-whisper
imageio-ffmpeg
